refactor catalog a little more
* consolidate NavBar component

* remove CoreLF, inject MUI theme directly, introduce Layout component

* flat routes
