--- catalog/app/components/SignIn/messages.js
@@ -1,9 +1,0 @@
-/* SignIn Messages */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  logIn: {
-    id: 'app.containers.UserMenu.logIn',
-    defaultMessage: 'Sign in',
-  },
-});
