--- catalog/app/components/SignIn/tests/index.test.js
@@ -1,10 +1,0 @@
-// import React from 'react';
-// import { shallow } from 'enzyme';
-
-// import SignIn from '../index';
-
-describe('<SignIn />', () => {
-  it('Expect to have unit tests specified', () => {
-    // expect(true).toEqual(false);
-  });
-});
