--- catalog/app/components/UserMenu/messages.js
@@ -1,13 +1,0 @@
-/* UserMenu Messages */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  logOut: {
-    id: 'app.containers.UserMenu.logOut',
-    defaultMessage: 'Sign out',
-  },
-  profile: {
-    id: 'app.containers.UserMenu.profile',
-    defaultMessage: 'Your profile',
-  },
-});
