--- catalog/app/components/UserMenu/tests/index.test.js
@@ -1,10 +1,0 @@
-// import React from 'react';
-// import { shallow } from 'enzyme';
-
-// import SignIn from '../index';
-
-describe('<UserMenu />', () => {
-  it('Expect to have unit tests specified', () => {
-    // expect(true).toEqual(false);
-  });
-});
