--- catalog/app/containers/AuthBar/messages.js
@@ -1,5 +1,0 @@
-/* AuthBar Messages */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-});
