+++ catalog/app/containers/HomePage/index.js
@@ -1,15 +1,16 @@
-/* HomePage */
-import React from 'react';
+import * as React from 'react';
 
 import CTA from 'components/CallToAction';
 import Feature from 'components/Feature';
+import Layout from 'components/Layout';
 import { BigSkip, Pad, UnPad } from 'components/LayoutHelpers';
 import QButton from 'components/QButton';
+import { composeComponent } from 'utils/reactTools';
 
 
-export default class HomePage extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
-  render() {
-    return (
+export default composeComponent('HomePage',
+  () => (
+    <Layout>
       <UnPad>
         <Feature
           header="T4 Web Catalog - Alpha Preview"
@@ -26,6 +27,5 @@ export default class HomePage extends React.PureComponent { // eslint-disable-li
         </Pad>
         <BigSkip />
       </UnPad>
-    );
-  }
-}
+    </Layout>
+  ));
